requirements.txt
confluent_kafka
pyspark
simplejson
python-dotenv
boto3
requests